3D-Object-Detection--3d-bounding-boxes--with-mediapipe-objectron-main/requirements.txt
mediapipe==0.10.1
opencv-python-headless==4.7.0.72
numpy
Pillow
matplotlib
streamlit
pandas
